+++ examples/bond/scripts/run/epoch1.js
@@ -55,7 +55,7 @@ exports.epoch1 = async function (deployer) {
     },
     // call to bond-dapp
     {
-      type: types.TransactionType.CallNoOpSSC,
+      type: types.TransactionType.CallApp,
       sign: types.SignType.SecretKey,
       fromAccount: account.elon,
       appID: appInfo.appID,