+++ packages/runtime/test/integration/updateApp.ts
@@ -62,7 +62,7 @@ describe("Algorand Smart Contracts - Update Application", function () {
     // now call the smart contract after updating approval program which checks for
     // global-key and local-key in state (which was set during the update from oldApprovalProgram)
     const noOpParams: types.AppCallsParam = {
-      type: types.TransactionType.CallNoOpSSC,
+      type: types.TransactionType.CallApp,
       sign: types.SignType.SecretKey,
       fromAccount: creator.account,
       appID: appID,